Validacion fecha eliminacion

+++ Schedules/BAL/Rules/ScheduleBAL.cs
@@ -14,6 +14,7 @@ namespace BAL
         private ScheduleValidation<Schedule> Validations { get; set; }
         private int Id { get; set; }
         private Schedule Schedule { get; set; }
+        const int HOURSTOCANCEL = 24;
 
         public ScheduleBAL()
         {
@@ -29,8 +30,14 @@ namespace BAL
 
         public void Delete()
         {
-            //TODO: validation needed before Delete it.
-            Repository.Delete(Schedule.Id);
+            if (IsValidToCancel(Schedule.Id, Schedule.Datebook))
+            {
+                Repository.Delete(Schedule.Id);
+            }
+            else
+            {
+                throw new ApplicationException("La cita no puede cancelarse.");
+            }
         }
 
         public object Get()
@@ -54,11 +61,35 @@ namespace BAL
             }
         }
 
+        /// <summary>
+        /// Validation schedule same day rule.
+        /// </summary>
+        /// <param name="id">Identify Pacient</param>
+        /// <param name="datebook">Datetime schedule</param>
+        /// <returns>True when process can continue.</returns>
         public bool IsValidedCreateSchedule(int id, DateTime datebook)
         {
             bool isValid = true;
             var schedulesDay = Validations.GetSchedulesSameDay(id, datebook);
             if (schedulesDay != null || datebook < System.DateTime.Now)
+            {
+                isValid = false;
+            }
+            return isValid;
+        }
+
+        /// <summary>
+        /// Validation schedule can to cancel.
+        /// </summary>
+        /// <param name="id">Id Schedule</param>
+        /// <param name="datebook">Datime schedule</param>
+        /// <returns>True when process can continue.</returns>
+        public bool IsValidToCancel(int id, DateTime datebook)
+        {
+            bool isValid = true;
+            
+            var schedules = Validations.GetScheduleToCancel(id, datebook);
+            if (schedules != null && schedules.Datebook.AddHours(-HOURSTOCANCEL) < DateTime.Now) 
             {
                 isValid = false;
             }
